+++ src/features/search/components/index.ts
@@ -1,2 +1,3 @@
 export { default as SearchForm } from './searchForm';
 export { default as SearchList } from './searchList';
+export { default as LatelySearch } from './latelySearch';
