+++ src/features/search/components/latelySearch/style.tsx
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled';
+
+export const WrapperTitle = styled.section`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+export const WrapperContent = styled.section`
+  display: flex;
+  align-items: center;
+  gap: 0 0.75rem;
+  overflow-x: scroll;
+`;
+
+export const KeywordBox = styled.div`
+  flex-shrink: 0;
+`;
