+++ src/pages/Search/SearchMain/index.tsx
@@ -1,55 +1,15 @@
-import { useNavigate, useOutletContext } from 'react-router-dom';
-import { CommonText, CommonButton, CommonTag } from '@/shared/components';
-import { Container, WrapperTitle, WrapperContent, KeywordBox } from './style';
-import { SearchList } from '@/features/search/components';
-import { useSearchedKeywordStorage } from '@/features/search/hooks';
-
-// TODO:input에 검색어가 없을시
-// 1. localstorage에 저장되어 있는 최근 검색어 보여주기 (O)
-// 2. 모두 지우기 기능 (O)
-// 3. 하나씩 지우기 기능 (O)
-// 4. 클릭시 이동 기능
-// 5. 추가하기 기능
+import { useOutletContext } from 'react-router-dom';
+import { Container } from './style';
+import { LatelySearch, SearchList } from '@/features/search/components';
 
 const SearchMain = () => {
   const keyword = useOutletContext<string>();
-
-  const navigate = useNavigate();
-
-  const [storageValue, setState] = useSearchedKeywordStorage('search');
 
   const isWord = !keyword || keyword.length <= 0;
 
   return isWord ? (
     <Container>
-      <WrapperTitle>
-        <CommonText type="normalInfo">최근 검색어</CommonText>
-        <CommonButton
-          type="xsText"
-          onClick={() => {
-            setState([]);
-          }}
-        >
-          모두 지우기
-        </CommonButton>
-      </WrapperTitle>
-      <WrapperContent>
-        {storageValue.length > 0 &&
-          storageValue.map((value: string, index: number) => (
-            <KeywordBox key={index}>
-              <CommonTag
-                type="search"
-                onClick={() => navigate('result')}
-                onDelete={() => {
-                  const filterData = storageValue.filter((keyword: string) => value !== keyword);
-                  setState([...filterData]);
-                }}
-              >
-                {value}
-              </CommonTag>
-            </KeywordBox>
-          ))}
-      </WrapperContent>
+      <LatelySearch />
     </Container>
   ) : (
     <SearchList keyword={keyword} />
