+++ src/pages/Search/SearchMain/style.tsx
@@ -6,20 +6,3 @@ export const Container = styled.article`
   flex-direction: column;
   gap: 1rem 0;
 `;
-
-export const WrapperTitle = styled.section`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-export const WrapperContent = styled.section`
-  display: flex;
-  align-items: center;
-  gap: 0 0.75rem;
-  overflow-x: scroll;
-`;
-
-export const KeywordBox = styled.div`
-  flex-shrink: 0;
-`;
